Tests/QUnit: Fix - Parents/modules/tests names.
ref: ED-2262
