+++ tests/qunit/tests/assets/dev/js/editor/data/globals/typography/component.spec.js
@@ -1,8 +1,6 @@
 import DocumentHelper from 'elementor/tests/qunit/tests/assets/dev/js/editor/document/helper';
 import * as commands from './commands/index.spec';
 
-jQuery( () => {
-	QUnit.module( 'Component: globals/typography', () => {
-		DocumentHelper.testCommands( commands );
-	} );
+QUnit.module( 'Component: globals/typography', () => {
+	DocumentHelper.testCommands( commands );
 } );
