+++ tests/qunit/tests/assets/dev/js/editor/document/repeater/component.spec.js
@@ -1,8 +1,6 @@
 import DocumentHelper from '../helper';
 import * as commands from './commands/index.spec.js';
 
-jQuery( () => {
-	QUnit.module( 'Component: document/repeater', () => {
-		DocumentHelper.testCommands( commands );
-	} );
+QUnit.module( 'Component: document/repeater', () => {
+	DocumentHelper.testCommands( commands );
 } );
