+++ tests/qunit/tests/assets/dev/js/editor/document/ui/component.spec.js
@@ -2,12 +2,10 @@ import DocumentHelper from '../helper';
 import ElementsHelper from '../elements/helper';
 import * as commands from './commands/index.spec.js';
 
-jQuery( () => {
-	QUnit.module( 'Component: document/ui', ( hooks ) => {
-		hooks.beforeEach = () => {
-			ElementsHelper.empty();
-		};
+QUnit.module( 'Component: document/ui', ( hooks ) => {
+	hooks.beforeEach = () => {
+		ElementsHelper.empty();
+	};
 
-		DocumentHelper.testCommands( commands );
-	} );
+	DocumentHelper.testCommands( commands );
 } );
